refactor: write 훅 생성 및 로직  분리
